+++ include/Instrument.h
@@ -29,7 +29,7 @@
 #include <QString>
 #include "export.h"
 #include "lmms_basics.h"
-#include "MemoryManager.h"
+#include "Memory.h"
 #include "MidiTime.h"
 #include "Plugin.h"
