+++ include/ValueBuffer.h
@@ -28,7 +28,7 @@
 
 #include <vector>
 
-#include "MemoryManager.h"
+#include "Memory.h"
 #include "export.h"
 
 class EXPORT ValueBuffer : public std::vector<float>
