+++ plugins/kicker/CMakeLists.txt
@@ -1,3 +1,3 @@
 INCLUDE(BuildPlugin)
 
-BUILD_PLUGIN(kicker kicker.cpp kicker.h MOCFILES kicker.h EMBEDDED_RESOURCES artwork.png logo.png)
+BUILD_PLUGIN(kicker kicker.cpp kicker.h KickerOsc.h MOCFILES kicker.h EMBEDDED_RESOURCES artwork.png logo.png)
