+++ plugins/nes/Nes.h
@@ -34,7 +34,7 @@
 #include "TempoSyncKnob.h"
 #include "NotePlayHandle.h"
 #include "PixmapButton.h"
-#include "MemoryManager.h"
+#include "Memory.h"
 
 
 #define makeknob( name, x, y, hint, unit, oname ) 		\
